requirements.txt
alabaster==1.0.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
docutils==0.21.2
fake-useragent==2.2.0
free_proxy==1.1.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20250416
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
roman-numerals-py==3.1.0
scholarly==1.7.11
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
